Builds Windows nounity and unity Builds Linux nounity and unity

+++ Code/Tools/AssetProcessor/native/AssetManager/SourceFileRelocator.cpp
@@ -699,7 +699,7 @@ Please note that only those seed files will get updated that are active for your
                 {
                     report.append(AZStd::string::format(
                         "SOURCEID: %" PRId64 ", CURRENT PATH: %s, NEW PATH: %s, CURRENT GUID: %s, NEW GUID: %s\n",
-                        static_cast<int64_t>(relocationInfo.m_sourceEntry.m_sourceID),
+                        relocationInfo.m_sourceEntry.m_sourceID,
                         relocationInfo.m_oldRelativePath.c_str(),
                         relocationInfo.m_newRelativePath.c_str(),
                         relocationInfo.m_sourceEntry.m_sourceGuid.ToString<AZStd::string>().c_str(),
@@ -709,7 +709,7 @@ Please note that only those seed files will get updated that are active for your
                 {
                     report.append(AZStd::string::format(
                         "SOURCEID: %" PRId64 ", CURRENT PATH: %s, CURRENT GUID: %s\n",
-                        static_cast<int64_t>(relocationInfo.m_sourceEntry.m_sourceID),
+                        relocationInfo.m_sourceEntry.m_sourceID,
                         relocationInfo.m_oldRelativePath.c_str(),
                         relocationInfo.m_sourceEntry.m_sourceGuid.ToString<AZStd::string>().c_str()));
                 }
